refactor(mcp): drain coworker-capability tool domain into a scoped pack (EP-8DC217EB BET-4)
Drain the coworker self-assessment / capability-needs domain
(get_my_coworker_profile, assess_my_capabilities,
submit_coworker_capability_need, list_my_capability_needs,
list_all_capability_needs) out of the mcp-tools.ts executeTool switch into a
scoped, lazily-dispatched pack. The profile/needs helpers
(requireCurrentCoworker, loadCoworkerProfile + its route-defined fallback,
parseCapabilityNeeds) were used only by these handlers and move with them;
now-unused imports are removed. Behaviour is preserved verbatim; grants mirror
agent-grants.ts TOOL_TO_GRANTS.

Process-Spine-Decision: EP-8DC217EB plan §4 BET-4 (kernel: drain-to-zero) — drain the coworker-capability tool domain into a scoped pack (registered via pack-registry.ts); ratchet baseline shrunk; behavior-preserving relocation.
