+++ apps/web/lib/mcp-tools.ts
@@ -50,25 +50,7 @@ import {
   isSandboxRecoveryAction,
 } from "@/lib/integrate/sandbox/sandbox-admin-types";
 import { getToolMarketplaceReadiness } from "@/lib/actions/tool-marketplace-readiness";
-import {
-  listCoworkerCapabilityNeeds,
-  submitCoworkerSelfAssessment,
-} from "@/lib/coworker-self-assessment/assessment-service";
 import { inferProviderIdFromRouteContext } from "@/lib/ai-provider-route-context";
-import {
-  COWORKER_ASSESSMENT_CONFIDENCE,
-  COWORKER_ASSESSMENT_VERDICTS,
-  COWORKER_CAPABILITY_NEED_KINDS,
-  COWORKER_CAPABILITY_NEED_SEVERITIES,
-  COWORKER_CAPABILITY_NEED_STATUSES,
-  type CoworkerAssessmentConfidence,
-  type CoworkerAssessmentVerdict,
-  type CoworkerCapabilityNeedInput,
-  type CoworkerCapabilityNeedKind,
-  type CoworkerCapabilityNeedSeverity,
-  type CoworkerCapabilityNeedStatus,
-} from "@/lib/coworker-self-assessment/types";
-import { ROUTE_AGENT_MAP_ENTRIES } from "@/lib/tak/agent-routing";
 // ─── Types ───────────────────────────────────────────────────────────────────
 export type BuildPhaseTag = "ideate" | "plan" | "build" | "review" | "ship";
 type ToolExecutionContext = {
@@ -3020,131 +3002,6 @@ export const PLATFORM_TOOLS: ToolDefinition[] = [
     },
   },
   {
-    name: "get_my_coworker_profile",
-    description: "Read the current coworker's registry identity, skills, grants, latest self-assessment, and route context. Use before answering whether the coworker has the tools/capabilities to do a job.",
-    inputSchema: {
-      type: "object",
-      properties: {},
-    },
-    requiredCapability: "view_platform",
-    executionMode: "immediate",
-    sideEffect: false,
-    buildPhases: ["ideate", "plan", "build", "review", "ship"],
-    annotations: {
-      readOnlyHint: true,
-      idempotentHint: true,
-    },
-  },
-  {
-    name: "assess_my_capabilities",
-    description: "Build current self-assessment context for the current coworker, including profile, skills, grants, marketplace readiness, and the response shape expected for capability needs.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        query: {
-          type: "string",
-          description: "Optional domain or need to focus readiness lookup, such as 'marketing publishing' or 'incident response'.",
-        },
-      },
-    },
-    requiredCapability: "view_platform",
-    executionMode: "immediate",
-    sideEffect: false,
-    buildPhases: ["ideate", "plan", "build", "review", "ship"],
-    annotations: {
-      readOnlyHint: true,
-      idempotentHint: true,
-    },
-  },
-  {
-    name: "submit_coworker_capability_need",
-    description: "Submit the current coworker's self-assessment and capability needs for human review. This creates CoworkerCapabilityNeed records, not direct BacklogItems.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        verdict: { type: "string", enum: [...COWORKER_ASSESSMENT_VERDICTS] },
-        confidence: { type: "string", enum: [...COWORKER_ASSESSMENT_CONFIDENCE] },
-        missionSummary: { type: "string", description: "How the coworker understands its mission/job in one or two sentences." },
-        capabilitySummary: { type: "string", description: "Short summary of what the coworker can do now and where it is limited." },
-        rawPayload: { type: "object", description: "Optional raw structured assessment payload for audit." },
-        needs: {
-          type: "array",
-          description: "Specific capability needs to submit for review.",
-          items: {
-            type: "object",
-            properties: {
-              kind: { type: "string", enum: [...COWORKER_CAPABILITY_NEED_KINDS] },
-              severity: { type: "string", enum: [...COWORKER_CAPABILITY_NEED_SEVERITIES] },
-              need: { type: "string" },
-              blocks: { type: "string" },
-              evidenceJson: { type: "object" },
-              readinessJson: { type: "object" },
-            },
-            required: ["kind", "severity", "need", "blocks"],
-          },
-        },
-      },
-      required: ["verdict", "confidence", "needs"],
-    },
-    requiredCapability: "view_platform",
-    executionMode: "immediate",
-    sideEffect: true,
-    buildPhases: ["ideate", "plan", "build", "review", "ship"],
-    annotations: {
-      destructiveHint: false,
-      idempotentHint: false,
-    },
-  },
-  {
-    name: "list_my_capability_needs",
-    description: "List capability needs previously submitted by the current coworker, optionally filtered by status, kind, or severity.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        status: { type: "string", enum: [...COWORKER_CAPABILITY_NEED_STATUSES] },
-        kind: { type: "string", enum: [...COWORKER_CAPABILITY_NEED_KINDS] },
-        severity: { type: "string", enum: [...COWORKER_CAPABILITY_NEED_SEVERITIES] },
-      },
-    },
-    requiredCapability: "view_platform",
-    executionMode: "immediate",
-    sideEffect: false,
-    buildPhases: ["ideate", "plan", "build", "review", "ship"],
-    annotations: {
-      readOnlyHint: true,
-      idempotentHint: true,
-    },
-  },
-  {
-    // BI-F9E7B780: governance-surface rollup. Same shape as
-    // list_my_capability_needs but UN-scoped from the calling agent —
-    // returns all coworker needs across all agents, plus filter-options
-    // and counts-by-status/severity/kind summaries that the existing
-    // /platform/ai/capability-needs admin page already renders. Lets
-    // taxonomy audits query governed capability evidence without direct
-    // SQL fallback.
-    name: "list_all_capability_needs",
-    description:
-      "List capability needs submitted by ALL coworkers (not scoped to the calling agent), optionally filtered by agentId, status, kind, or severity. Returns the rich review shape used by the /platform/ai/capability-needs admin page: summary counts by status/severity/kind, filter-option enums, and the full need list. Read-only governance surface for taxonomy audits.",
-    inputSchema: {
-      type: "object",
-      properties: {
-        agentId: { type: "string", description: "Optional — filter to one coworker's needs. Omit to return all." },
-        status: { type: "string", enum: [...COWORKER_CAPABILITY_NEED_STATUSES] },
-        kind: { type: "string", enum: [...COWORKER_CAPABILITY_NEED_KINDS] },
-        severity: { type: "string", enum: [...COWORKER_CAPABILITY_NEED_SEVERITIES] },
-      },
-    },
-    requiredCapability: "view_platform",
-    executionMode: "immediate",
-    sideEffect: false,
-    buildPhases: ["ideate", "plan", "build", "review", "ship"],
-    annotations: {
-      readOnlyHint: true,
-      idempotentHint: true,
-    },
-  },
-  {
     name: "prefill_onboarding_wizard",
     description: "Pre-fill the regulation onboarding wizard with AI-drafted data. Stores a draft and returns the wizard URL for human review.",
     inputSchema: {
@@ -3494,142 +3351,6 @@ async function updateBuildHappyPathState(
     where: { buildId: resolvedBuildId },
     data: { plan: mergedPlan as import("@dpf/db").Prisma.InputJsonValue },
   });
-}
-
-function requireCurrentCoworker(context?: { agentId?: string }): string {
-  const agentId = context?.agentId?.trim();
-  if (!agentId) {
-    throw new Error("A current coworker agentId is required for this tool.");
-  }
-  return agentId;
-}
-
-function routeValueStream(capability: CapabilityKey | null): string {
-  return capability?.replace(/^(view|manage)_/, "") || "cross-cutting";
-}
-
-async function loadRouteDefinedCoworkerProfile(agentId: string, routeContext?: string | null) {
-  const routeEntry = ROUTE_AGENT_MAP_ENTRIES.find(([, entry]) => entry.agentId === agentId)?.[1] ?? null;
-  if (!routeEntry) return null;
-
-  const { getAgentToolGrants } = await import("@/lib/tak/agent-grants");
-  const grants = getAgentToolGrants(agentId) ?? ["registry_read"];
-
-  return {
-    profile: {
-      agentId: routeEntry.agentId,
-      slugId: routeEntry.agentId,
-      name: routeEntry.agentName,
-      tier: routeEntry.sensitivity === "restricted" ? 1 : 2,
-      type: "specialist",
-      description: routeEntry.agentDescription,
-      valueStream: routeValueStream(routeEntry.capability),
-      it4itSections: [],
-      lifecycleStage: "production",
-      hitlTierDefault: routeEntry.sensitivity === "public" ? 3 : 1,
-      humanSupervisorId: null,
-      escalatesTo: null,
-      delegatesTo: [],
-      routeContext: routeContext ?? null,
-      grants,
-      skills: routeEntry.skills.map((skill, index) => ({
-        label: skill.label,
-        description: skill.description,
-        capability: skill.capability,
-        taskType: skill.taskType ?? "conversation",
-        sortOrder: index,
-      })),
-    },
-    latestAssessment: null,
-  };
-}
-
-async function loadCoworkerProfile(agentId: string, routeContext?: string | null) {
-  const agent = await prisma.agent.findFirst({
-    where: { OR: [{ agentId }, { slugId: agentId }] },
-    include: {
-      skills: {
-        orderBy: { sortOrder: "asc" },
-        select: {
-          label: true,
-          description: true,
-          capability: true,
-          taskType: true,
-          sortOrder: true,
-        },
-      },
-      toolGrants: {
-        select: { grantKey: true },
-        orderBy: { grantKey: "asc" },
-      },
-      coworkerAssessments: {
-        orderBy: { createdAt: "desc" },
-        take: 1,
-        select: {
-          assessmentId: true,
-          verdict: true,
-          confidence: true,
-          createdAt: true,
-        },
-      },
-    },
-  });
-
-  if (!agent) {
-    const routeDefinedProfile = await loadRouteDefinedCoworkerProfile(agentId, routeContext);
-    if (routeDefinedProfile) {
-      return routeDefinedProfile;
-    }
-
-    throw new Error(`Coworker ${agentId} was not found.`);
-  }
-
-  const latestAssessment = agent.coworkerAssessments[0] ?? null;
-  return {
-    profile: {
-      agentId: agent.agentId,
-      slugId: agent.slugId,
-      name: agent.name,
-      tier: agent.tier,
-      type: agent.type,
-      description: agent.description,
-      valueStream: agent.valueStream,
-      it4itSections: agent.it4itSections,
-      lifecycleStage: agent.lifecycleStage,
-      hitlTierDefault: agent.hitlTierDefault,
-      humanSupervisorId: agent.humanSupervisorId,
-      escalatesTo: agent.escalatesTo,
-      delegatesTo: agent.delegatesTo,
-      routeContext: routeContext ?? null,
-      grants: agent.toolGrants.map((grant) => grant.grantKey),
-      skills: agent.skills,
-    },
-    latestAssessment,
-  };
-}
-
-function parseCapabilityNeeds(rawNeeds: unknown): CoworkerCapabilityNeedInput[] {
-  if (!Array.isArray(rawNeeds)) return [];
-
-  return rawNeeds
-    .filter((need): need is Record<string, unknown> => need != null && typeof need === "object")
-    .filter((need): need is Record<string, unknown> =>
-      COWORKER_CAPABILITY_NEED_KINDS.includes(need.kind as CoworkerCapabilityNeedKind))
-    .map((need) => ({
-      kind: need.kind as CoworkerCapabilityNeedKind,
-      severity: COWORKER_CAPABILITY_NEED_SEVERITIES.includes(need.severity as CoworkerCapabilityNeedSeverity)
-        ? need.severity as CoworkerCapabilityNeedSeverity
-        : "important",
-      need: String(need.need ?? "").trim(),
-      blocks: String(need.blocks ?? "").trim(),
-      evidenceJson: need.evidenceJson != null && typeof need.evidenceJson === "object"
-        ? need.evidenceJson as Record<string, unknown>
-        : undefined,
-      readinessJson: need.readinessJson != null && typeof need.readinessJson === "object"
-        ? need.readinessJson as Record<string, unknown>
-        : undefined,
-    }))
-    .filter((need) => need.need.length > 0 && need.blocks.length > 0);
 }
 
 function redactFunctionalFailureText(text: string): string {
@@ -11675,154 +11396,6 @@ export async function executeTool(
         data: {
           summary: readiness.summary,
           entries: readiness.entries,
-        },
-      };
-    }
-
-    case "get_my_coworker_profile": {
-      const agentId = requireCurrentCoworker(context);
-      const profile = await loadCoworkerProfile(agentId, context?.routeContext ?? null);
-      return {
-        success: true,
-        message: `Loaded coworker profile for ${profile.profile.name}.`,
-        data: profile,
-      };
-    }
-
-    case "assess_my_capabilities": {
-      const agentId = requireCurrentCoworker(context);
-      const profile = await loadCoworkerProfile(agentId, context?.routeContext ?? null);
-      const readiness = await getToolMarketplaceReadiness({
-        query: typeof params["query"] === "string" ? params["query"] : undefined,
-        agentId,
-        limit: 12,
-      });
-
-      return {
-        success: true,
-        message: `Prepared self-assessment context for ${profile.profile.name}.`,
-        data: {
-          ...profile,
-          readiness,
-          responseShape: {
-            verdict: [...COWORKER_ASSESSMENT_VERDICTS],
-            confidence: [...COWORKER_ASSESSMENT_CONFIDENCE],
-            needs: {
-              kind: [...COWORKER_CAPABILITY_NEED_KINDS],
-              severity: [...COWORKER_CAPABILITY_NEED_SEVERITIES],
-              requiredFields: ["kind", "severity", "need", "blocks"],
-            },
-          },
-          instruction:
-            "Assess your mission against your current skills, grants, and readiness. Submit durable gaps with submit_coworker_capability_need instead of creating backlog items directly.",
-        },
-      };
-    }
-
-    case "submit_coworker_capability_need": {
-      const agentId = requireCurrentCoworker(context);
-      const needs = parseCapabilityNeeds(params["needs"]);
-      if (needs.length === 0) {
-        return {
-          success: false,
-          error: "No valid capability needs supplied",
-          message: "Submit at least one need with kind, severity, need, and blocks.",
-        };
-      }
-
-      const verdict = COWORKER_ASSESSMENT_VERDICTS.includes(params["verdict"] as CoworkerAssessmentVerdict)
-        ? params["verdict"] as CoworkerAssessmentVerdict
-        : "gaps";
-      const confidence = COWORKER_ASSESSMENT_CONFIDENCE.includes(params["confidence"] as CoworkerAssessmentConfidence)
-        ? params["confidence"] as CoworkerAssessmentConfidence
-        : "medium";
-
-      const result = await submitCoworkerSelfAssessment({
-        agentId,
-        trigger: "tool-call",
-        routeContext: context?.routeContext ?? null,
-        verdict,
-        confidence,
-        missionSummary: typeof params["missionSummary"] === "string" ? params["missionSummary"] : null,
-        capabilitySummary: typeof params["capabilitySummary"] === "string" ? params["capabilitySummary"] : null,
-        rawPayload: params["rawPayload"] != null && typeof params["rawPayload"] === "object"
-          ? params["rawPayload"] as Record<string, unknown>
-          : { toolName },
-        needs,
-      });
-
-      return {
-        success: true,
-        entityId: result.assessmentId,
-        message: `Submitted ${result.needIds.length} capability need${result.needIds.length === 1 ? "" : "s"} and filed ${result.backlogItemIds?.length ?? 0} to the backlog for triage.`,
-        data: result,
-      };
-    }
-
-    case "list_my_capability_needs": {
-      const agentId = requireCurrentCoworker(context);
-      const status = COWORKER_CAPABILITY_NEED_STATUSES.includes(params["status"] as CoworkerCapabilityNeedStatus)
-        ? params["status"] as CoworkerCapabilityNeedStatus
-        : undefined;
-      const kind = COWORKER_CAPABILITY_NEED_KINDS.includes(params["kind"] as CoworkerCapabilityNeedKind)
-        ? params["kind"] as CoworkerCapabilityNeedKind
-        : undefined;
-      const severity = COWORKER_CAPABILITY_NEED_SEVERITIES.includes(params["severity"] as CoworkerCapabilityNeedSeverity)
-        ? params["severity"] as CoworkerCapabilityNeedSeverity
-        : undefined;
-      const needs = await listCoworkerCapabilityNeeds({ agentId, status, kind, severity });
-
-      return {
-        success: true,
-        message: `Found ${needs.length} capability need${needs.length === 1 ? "" : "s"}.`,
-        data: { needs },
-      };
-    }
-
-    case "list_all_capability_needs": {
-      // BI-F9E7B780 — governance-surface rollup, not scoped to the calling
-      // coworker. Reuses the same review-service helper that powers the
-      // /platform/ai/capability-needs admin page so the MCP surface and the
-      // operator UI agree on shape (summary counts, filter-option enums,
-      // full need list).
-      const { getCoworkerCapabilityNeedReview } = await import(
-        "@/lib/coworker-self-assessment/review-service"
-      );
-      const agentId = typeof params["agentId"] === "string" && params["agentId"].length > 0
-        ? params["agentId"]
-        : undefined;
-      const status = COWORKER_CAPABILITY_NEED_STATUSES.includes(
-        params["status"] as CoworkerCapabilityNeedStatus,
-      )
-        ? (params["status"] as CoworkerCapabilityNeedStatus)
-        : undefined;
-      const kind = COWORKER_CAPABILITY_NEED_KINDS.includes(
-        params["kind"] as CoworkerCapabilityNeedKind,
-      )
-        ? (params["kind"] as CoworkerCapabilityNeedKind)
-        : undefined;
-      const severity = COWORKER_CAPABILITY_NEED_SEVERITIES.includes(
-        params["severity"] as CoworkerCapabilityNeedSeverity,
-      )
-        ? (params["severity"] as CoworkerCapabilityNeedSeverity)
-        : undefined;
-
-      const review = await getCoworkerCapabilityNeedReview({
-        agentId,
-        status,
-        kind,
-        severity,
-      });
-
-      return {
-        success: true,
-        message: `Found ${review.summary.total} capability need${
-          review.summary.total === 1 ? "" : "s"
-        } across all coworkers.`,
-        data: {
-          summary: review.summary,
-          filterOptions: review.filterOptions,
-          needs: review.needs,
         },
       };
     }
