+++ apps/web/lib/mcp/pack-registry.ts
@@ -55,6 +55,7 @@ import { crmSalesPipelinePack } from "./packs/crm-sales-pipeline-pack";
 import { platformUpdatePack } from "./packs/platform-update-pack";
 import { wikiOverlayPack } from "./packs/wiki-overlay-pack";
 import { licensingPack } from "./packs/licensing-pack";
+import { coworkerCapabilityPack } from "./packs/coworker-capability-pack";
 
 export const TOOL_PACK_REGISTRY = composeToolPacks([
   deliberationSiemPack,
@@ -96,4 +97,5 @@ export const TOOL_PACK_REGISTRY = composeToolPacks([
   platformUpdatePack,
   wikiOverlayPack,
   licensingPack,
+  coworkerCapabilityPack,
 ]);
