+++ scripts/mcp-tool-pack-baseline.json
@@ -15,7 +15,6 @@
   "assess_archetype_refinement",
   "assess_complexity",
   "assess_contribution",
-  "assess_my_capabilities",
   "cancel_thread",
   "check_deployment_windows",
   "check_sandbox",
@@ -57,7 +56,6 @@
   "get_code_graph_freshness",
   "get_finance_period_summary",
   "get_marketing_summary",
-  "get_my_coworker_profile",
   "get_next_recommended_work",
   "get_release_status",
   "get_thread_result",
@@ -66,11 +64,9 @@
   "inspect_build_code_impact",
   "iterate_sandbox",
   "link_backlog_item_to_epic",
-  "list_all_capability_needs",
   "list_backlog_items",
   "list_build_activity_since",
   "list_epics",
-  "list_my_capability_needs",
   "list_patch_posture",
   "list_project_directory",
   "list_sandbox_files",
@@ -136,7 +132,6 @@
   "start_ideate_research",
   "start_sandbox",
   "start_scout_research",
-  "submit_coworker_capability_need",
   "submit_feedback",
   "suggest_campaign_ideas",
   "suggest_taxonomy_placement",
